Self-Build API
API was created with a skeleton from Back-End course but most of the things were removed. It was created shortly after the project creation but forgot to add it to the repo. Future changes will be reflected. API endpoints for the user and training programs are completed with the most basic and unsecure way of implementing it just for them to work.